Rovio Work!
Implemented all changes and logic Rovio sent to me. Untested, but I have a good feeling about this!
Only thing is that we may need to play with different pipeline tunings to get it to work as well as it should in different lightings, but that shouldn't be too difficult now that our Pipelines are so much simpler to work with (bye bye RectData!)

Co-Authored-By: rovio-chen <122228076+rovio-chen@users.noreply.github.com>